Pair Programming: Irina+Vlad

+++ TramsA6/Controllers/TransportMeanController.cs
@@ -1,22 +1,27 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using Domain.Entities;
 using Domain.Interfaces;
 using EnsureThat;
 using Microsoft.AspNetCore.Mvc;
+using TramsA6.DTOS.CommentModels;
 using TramsA6.DTOS.TransportMeanModels;
 
 namespace TramsA6.Controllers
 {
-    [Route("api/[controller]")]
+    [Route("api/MeansOfTransport")]
     public class TransportMeanController : Controller
     {
         private readonly ITransportMeanRepository _repository;
+        private readonly IUserRepository _userRepository;
 
-        public TransportMeanController(ITransportMeanRepository repository)
+        public TransportMeanController(ITransportMeanRepository repository, IUserRepository userRepository)
         {
             Ensure.That(repository).IsNotNull();
+            Ensure.That(userRepository).IsNotNull();
             _repository = repository;
+            _userRepository = userRepository;
         }
 
         [HttpGet]
@@ -40,6 +45,8 @@ namespace TramsA6.Controllers
         [HttpGet("{id}", Name = "GetMeanOfTransportById")]
         public IActionResult GetMeanOfTransportById(Guid id)
         {
+            if (id.Equals(Guid.Empty))
+                return BadRequest();
             var meanOfTransport = _repository.GetById(id);
             if (meanOfTransport == null)
                 return NotFound();
@@ -49,6 +56,8 @@ namespace TramsA6.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(Guid id)
         {
+            if (id.Equals(Guid.Empty))
+                return BadRequest();
             var status = _repository.Delete(id);
             if (!status)
                 return NotFound();
@@ -58,6 +67,8 @@ namespace TramsA6.Controllers
         [HttpPut("{id}")]
         public IActionResult Put(Guid id, [FromBody] UpdateTransportMeanDTO updateTransportMeanDto)
         {
+            if (id.Equals(Guid.Empty))
+                return BadRequest();
             if (updateTransportMeanDto == null)
                 return BadRequest();
             var meanOfTransport = _repository.GetById(id);
@@ -69,6 +80,23 @@ namespace TramsA6.Controllers
             return NoContent();
         }
 
-       //todo: addComment to mean of transport functionality
+
+        [HttpPut("{idMeanOfTransport}/comment")]
+        public IActionResult AddCommentToMeanOfTransport(Guid idMeanOfTransport, [FromBody] CreateCommentDTO comment)
+        {
+            if (idMeanOfTransport.Equals(Guid.Empty))
+                return BadRequest();
+            if (comment == null)
+                return BadRequest();
+            var meanOfTransport = _repository.GetById(idMeanOfTransport);
+            if (meanOfTransport == null)
+                return NotFound();
+            //todo: automapper
+            Comment com = Comment.Create(meanOfTransport, _userRepository.GetById(comment.User), DateTime.Now,
+                comment.Text, 0, 0);
+
+            meanOfTransport.Comments.Append(com);
+            return NoContent();
+        }
     }
 }
